Switch settings and preferences managers to use threading.Lock
