+++ backend_py/pyproject.toml
@@ -3,6 +3,9 @@ line-length = 88
 indent-width = 4
 exclude = ["v4l2.py"] # exclude v4l2.py, since it's included as a lib directly
 
+[pycodestyle]
+max_line_length = 88
+
 [tool.ruff.lint]
 select = [
   # pycodestyle
