+++ backend_py/src/services/cameras/__init__.py
@@ -1,7 +1,7 @@
+from ..preferences.settings_manager import SettingsManager
 from .device_manager import DeviceManager
 from .device_utils import find_device_with_bus_info, list_diff
 from .pwm import SerialPWMController
-from .settings import SettingsManager
 
 __all__ = [
     "DeviceManager",
