+++ backend_py/src/services/preferences/__init__.py
@@ -1,3 +1,4 @@
 from .preferences_manager import PreferencesManager
+from .settings_manager import SettingsManager
 
-__all__ = ["PreferencesManager"]
+__all__ = ["PreferencesManager", "SettingsManager"]
